requirements.txt
# LangChain Core
langchain
langchain-core



# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub
sentence-transformers
langchain-chroma
langchain-community
langchain_groq

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

#vector database
chromadb

pypdf
streamlit
